Updated RecipeActivity
Made some changes to display the recipe name/ingredient/instructions in one window. Will need to change for better visual and organization. Prefer to have names on one screen so user can tap the name and load the ingredients/instructions on another screen. This prevents having an error when a text field is blank (null), too.